requirements.txt
aiosignal==1.3.1
alembic==1.10.4
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
psycopg2-binary==2.9.6
pydantic==1.10.7
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
yarl==1.9.1
